+++ ThirdPartyAdapters/bidmachine/bidmachine/src/main/kotlin/com/google/ads/mediation/bidmachine/BidMachineMediationAdapter.kt
@@ -21,6 +21,7 @@ import com.google.ads.mediation.bidmachine.BidMachineBannerAd.Companion.mapAdSiz
 import com.google.android.gms.ads.AdError
 import com.google.android.gms.ads.AdFormat
 import com.google.android.gms.ads.AdSize
+import com.google.android.gms.ads.AgeRestrictedTreatment
 import com.google.android.gms.ads.MobileAds.getRequestConfiguration
 import com.google.android.gms.ads.RequestConfiguration
 import com.google.android.gms.ads.VersionInfo
@@ -303,9 +304,11 @@ constructor(private val mediationUtils: MediationUtilsWrapper = MediationUtilsWr
   private fun configureBidMachinePrivacy() {
     val tagForChildDirected = getRequestConfiguration().tagForChildDirectedTreatment
     val tagForUnderAgeOfConsent = getRequestConfiguration().tagForUnderAgeOfConsent
+    val ageRestrictedTreatment = getRequestConfiguration().ageRestrictedTreatment
     if (
       tagForChildDirected == RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_TRUE ||
-        tagForUnderAgeOfConsent == RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_TRUE
+        tagForUnderAgeOfConsent == RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_TRUE ||
+        ageRestrictedTreatment == AgeRestrictedTreatment.CHILD
     ) {
       BidMachine.setCoppa(true)
     } else if (